Feat: StringCalculator생성 및 테스트 코드 추가

+++ src/main/java/calculator/Calculator.java
@@ -1,21 +1,21 @@
 package calculator;
 
 public class Calculator {
-    private final Long operand1;
+    private final Double operand1;
     private final String operator;
-    private final Long operand2;
+    private final Double operand2;
 
     public Calculator() {
-        this(0L, "+", 0L);
+        this(0.0, "+", 0.0);
     }
 
-    public Calculator(Long operand1, String operator, Long operand2) {
+    public Calculator(Double operand1, String operator, Double operand2) {
         this.operand1 = operand1;
         this.operator = operator;
         this.operand2 = operand2;
     }
 
-    public Long calculate(Long operand1, String operator, Long operand2) {
+    public Double calculate(Double operand1, String operator, Double operand2) {
         if ("+".equals(operator)) {
             return operand1 + operand2;
         }
@@ -28,6 +28,6 @@ public class Calculator {
         if ("/".equals(operator)) {
             return operand1 / operand2;
         }
-        return 0L;
+        return 0.0;
     }
 }
